+++ src/card/cards/combat_constraints.rs
@@ -16,3 +16,4 @@ pub const XENIC_POLTERGEIST: CardDefinitionId = CardDefinitionId(1815);
 pub const IFH_BIFF_EFREET: CardDefinitionId = CardDefinitionId(1816);
 pub const ARMAGEDDON_CLOCK: CardDefinitionId = CardDefinitionId(1817);
 pub const OSAI_VULTURES: CardDefinitionId = CardDefinitionId(1818);
+pub const COCOON: CardDefinitionId = CardDefinitionId(1819);